+++ apps/dev-docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # dev-docs
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @brightweblabs/ui@1.4.2
+
 ## 0.0.9
 
 ### Patch Changes